+++ just_audio_platform_interface/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.2.2
+
+* Add setAllowsExternalPlayback on iOS/macOS.
+* Support index-based seeking on Android/iOS/macOS.
+
 ## 4.2.1
 
 * Update minimum flutter version to 3.0.